requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Lenovo\Documents\Python Codes\Python Files\Major Projects\binod-bot\discord_bot.py: 12
discord.py == 1.5.0

# C:\Users\Lenovo\Documents\Python Codes\Python Files\Major Projects\binod-bot\discord_bot.py: 12
requests == 2.21.0

# C:\Users\Lenovo\Documents\Python Codes\Python Files\Major Projects\binod-bot\discord_bot.py: 12
wikipedia == 1.4.0
